requirements.txt
amqp==5.1.1
aniso8601==7.0.0
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
billiard==4.1.0
boto3==1.28.28
botocore==1.31.28
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1

constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
daphne==4.0.0
dateparser==1.2.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2
django-admin-rangefilter==0.10.0
django-ckeditor==5.9.0
django-cors-headers==4.2.0
django-environ==0.10.0
django-filter==23.2
django-import-export==3.2.0
django-jazzmin==2.6.0
django-jet==1.0.8
django-jet-reboot==1.3.5
django-js-asset==2.2.0
django-model-utils==4.3.1
django-modeltranslation==0.18.11
django-rangefilter==0.9.1
django-scheduler==0.10.1
django-smart-selects==1.6.0
django-storages==1.13.2
django-suit==2.0a2
django-user-visit==0.4
djangorestframework==3.13.0
djangorestframework-simplejwt==5.3.0
docutils==0.20.1
docx==0.2.4
drf-yasg==1.21.7
drf-yasg2==1.19.4
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
executing==1.2.0
fdfgen==0.16.1
gevent==23.7.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==2.0.2
gunicorn==21.2.0
hyperlink==21.0.0
icalendar==5.0.11
idna==3.4
import-export==0.3.1
incremental==22.10.0
inflection==0.5.1

ipython-genutils==0.2.0
isort==5.12.0
itypes==1.2.0
jedi==0.19.0

Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.1
lxml==4.9.3
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
msgpack==1.0.5
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
parso==0.8.3
pdf2image==1.16.3
pdfrw==0.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
promise==2.3

psutil==5.9.7
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
regex==2023.12.25
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
Rx==1.6.3
s3transfer==0.6.2
schedule==1.2.1
sentry-sdk==1.29.2
service-identity==23.1.0
six==1.16.0
social-auth-core==4.5.1
SQLAlchemy==1.4.50
sqlparse==0.4.4
stack-data==0.6.2
tablib==3.5.0
traitlets==5.14.0
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.18
user-agents==2.2.0
uWSGI==2.0.22
vine==5.0.0
wcwidth==0.1.7
xlrd==2.0.1
xlwt==1.3.0